\name{addTimeStamp}
\alias{addTimeStamp}
\title{Adds timestamp to prompt}
\usage{
  addTimeStamp()
}
\description{
  Adds timestamp to prompt.  This will add the current time
  to the *beginning* of the prompt and leave your old
  prompt after the time stamp.  If you want to remove the
  timestamp and have it stop updating you can call
  \code{\link[timestamp]{removeTimeStamp}}.
}

