% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_falses_after_the_first}
\alias{assign_falses_after_the_first}
\title{Assign FALSE Values After First Encountered FALSE}
\usage{
assign_falses_after_the_first(pass_list, to_be_ignored)
}
\arguments{
\item{pass_list}{A LogicalList where TRUE represents valid entries and FALSE represents invalid entries.}

\item{to_be_ignored}{A LogicalList indicating positions that should be ignored when determining the first FALSE.}
}
\value{
A LogicalList where elements after the first detected FALSE remain FALSE.
}
\description{
Modifies a logical list such that all values after the first encountered FALSE are also set to FALSE.
}
