% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_seqlevels}
\alias{clean_seqlevels}
\title{Clean Sequence Levels in a GRanges Object}
\usage{
clean_seqlevels(x)
}
\arguments{
\item{x}{A GRanges object.}
}
\value{
A GRanges object with unused sequence levels removed.
}
\description{
Removes unused sequence levels from a GRanges object.
}
