% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_flanking_gap_length.R
\name{compute_flanking_gap_length}
\alias{compute_flanking_gap_length}
\title{Compute Flanking Gap Lengths of Loci That Do Not Have The Insertion}
\usage{
compute_flanking_gap_length(missings, ref)
}
\arguments{
\item{missings}{A `GRanges` object representing missing regions in the genome.}

\item{ref}{A `GRanges` object representing the reference TE annotations.}
}
\value{
A list containing:
  - `up`: A matrix of upstream gap lengths grouped by species.
  - `dn`: A matrix of downstream gap lengths grouped by species.
}
\description{
This function calculates the upstream and downstream gap lengths of locs that do not have the TE insertion
identifies flanking regions based on overlaps and extracts gaps using a logic similar to `setdiff` in `GenomicRanges`.
}
