% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{empty_metadata}
\alias{empty_metadata}
\title{Remove Metadata from a GRanges Object}
\usage{
empty_metadata(x)
}
\arguments{
\item{x}{A GRanges object.}
}
\value{
A GRanges object with metadata removed.
}
\description{
Removes metadata columns from a GRanges object.
}
