% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unaligned_regions.R
\name{extract_unaligned_regions_by_species}
\alias{extract_unaligned_regions_by_species}
\title{Identify Unaligned Regions from MAF and FASTA Files}
\usage{
extract_unaligned_regions_by_species(
  fa,
  maf,
  ref_genome = "hg38",
  folder,
  min_len = 10L
)
}
\arguments{
\item{fa}{Character. Path to the input FASTA file.}

\item{maf}{Character. Path to the input MAF file (only the first 3 lines are used).}

\item{ref_genome}{Character. Reference genome identifier (default is \code{"hg38"}).}

\item{min_len}{Integer. Minimum length (in base pairs) of the merged region to be saved (default is \code{10L}).}

\item{out_folder}{Character. Output directory to save merged region results.}
}
\value{
Saves a merged \code{.rds} file of \code{GRanges} objects representing the gaps and missing regions.
}
\description{
This function reads a multiple alignment format (MAF) file and a corresponding
FASTA file to identify regions that are missing (\code{*}) or gaps (\code{-}) in the alignment,
relative to a reference genome. It returns merged genomic regions meeting a
minimum length requirement.
}
\examples{
\dontrun{
get_unaligned_regions("example.fa", "example.maf", "hg38", "output_folder")
}
}
