% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_element}
\alias{first_element}
\title{Extract First Element from a List}
\usage{
first_element(x, invert = FALSE)
}
\arguments{
\item{x}{A list of elements.}

\item{invert}{If TRUE, returns all but the first element.}
}
\value{
A list with the first element extracted (or removed if invert = TRUE).
}
\description{
Extracts the first element of each sub-list within a list.
}
