% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coverage_by_clades.R
\name{get_coverage_by_clades}
\alias{get_coverage_by_clades}
\title{Compute Alignment Coverage by Clades}
\usage{
get_coverage_by_clades(rmsk_range, sister_clades, folder, chrom = NULL)
}
\arguments{
\item{rmsk_range}{A GRanges object representing repeatMasker regions.}

\item{sister_clades}{A list of sister-clades.}

\item{folder}{A character string specifying the directory where output files will be stored.}

\item{chrom}{A character value specifying the chromosome of interest (default:NULL).}
}
\description{
This function calculates the coverage of transposable element sequences across sister clades.
}
