% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timepoint.R
\name{get_timepoint}
\alias{get_timepoint}
\title{Estimates the insertion timepoint}
\usage{
get_timepoint(folder, consider_missing_loci = TRUE, chrom = NULL)
}
\arguments{
\item{folder}{A character string specifying the directory containing required data files.}

\item{consider_missing_loci}{A logical value indicating whether to account for missing loci (default: TRUE).}

\item{chrom}{A character string specifying the chromosome of interest.}
}
\value{
Saves the computed timepoint in `timepoint` in the specified `folder` directory.
}
\description{
This function estimates the insertion timepoint of sequences based on
alignment coverage across sister clades, with an optional adjustment for missing loci.
}
