% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timepoint.R
\name{get_timepoint_by_chrom}
\alias{get_timepoint_by_chrom}
\title{Infer Insertion Timepoints by Chromosome}
\usage{
get_timepoint_by_chrom(folder, chrom, consider_missing_loci = TRUE)
}
\arguments{
\item{folder}{A character string specifying the directory containing required data files.}

\item{chrom}{A character string specifying the chromosome of interest.}

\item{consider_missing_loci}{A logical value indicating whether to account for missing loci (default: TRUE).}
}
\value{
Saves the computed timepoint in `timepoint` in the specified `folder` directory.
}
\description{
Infer Insertion Timepoints by Chromosome
}
