% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rmsk.R
\name{import_rmsk}
\alias{import_rmsk}
\title{Import and Process RepeatMasker Output}
\usage{
import_rmsk(rmsk_file, folder = NULL)
}
\arguments{
\item{rmsk_file}{A character string specifying the path to the RepeatMasker output file.}

\item{folder}{An optional character string specifying the directory to save the processed data. Default is NULL.}
}
\value{
A `GRanges` object containing processed repeat annotations with class, family, and repeat coordinates.
}
\description{
This function reads a RepeatMasker output file and processes it into a
`GRanges` object with metadata, organizing repeat elements by their IDs.
}
