% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{last_element}
\alias{last_element}
\title{Extract Last Element from a List}
\usage{
last_element(x, invert = FALSE)
}
\arguments{
\item{x}{A list of elements.}

\item{invert}{If TRUE, returns all but the last element.}
}
\value{
A list with the last element extracted (or removed if invert = TRUE).
}
\description{
Extracts the last element of each sub-list within a list.
}
