% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mat2list}
\alias{mat2list}
\title{Convert a Matrix to a List}
\usage{
mat2list(x, by_col = FALSE)
}
\arguments{
\item{x}{A matrix.}

\item{by_col}{If TRUE, converts by columns; otherwise, converts by rows.}
}
\value{
A list representation of the matrix.
}
\description{
Converts a matrix into a list, by rows or by columns.
}
