% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_timepoint.R
\name{remove_if_before_na}
\alias{remove_if_before_na}
\title{Remove Values Before NA Timepoints}
\usage{
remove_if_before_na(tp, mat)
}
\arguments{
\item{tp}{A numeric vector of predicted timepoints.}

\item{mat}{A matrix indicating missing values.}
}
\value{
A numeric vector where invalid timepoints are set to NA.
}
\description{
Removes predicted timepoints if they occur before missing data.
}
