\name{removeTimeStamp}
\alias{removeTimeStamp}
\title{Removes the time stamp}
\usage{
  removeTimeStamp()
}
\description{
  Removes the time stamp from the prompt and resets to the
  old prompt. If you want to add the timestamp back again
  just make a call to
  \code{\link[timestamp]{addTimeStamp}}.
}

