% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_timepoint.R
\name{timepoint_is_confi}
\alias{timepoint_is_confi}
\title{Check Timepoint Confidence}
\usage{
timepoint_is_confi(tp, mat, na_is_confi = TRUE)
}
\arguments{
\item{tp}{A numeric vector of predicted timepoints.}

\item{mat}{A matrix indicating missing data.}

\item{na_is_confi}{Logical flag (default: TRUE) indicating whether NA values should be considered confident.}
}
\value{
A logical vector indicating whether each timepoint is confident.
}
\description{
Determines if the predicted timepoint is reliable.
}
