% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tree.R
\name{update_tree}
\alias{update_tree}
\title{Update Phylogenetic Tree with New Sister Clades}
\usage{
update_tree(tree, new_sister_clades)
}
\arguments{
\item{tree}{A `treedata` object representing the phylogenetic tree.}

\item{new_sister_clades}{A list where each element represents a sister clade,}
}
\value{
A `treedata` object with updated metadata for sister clade groups.
}
\description{
Update Phylogenetic Tree with New Sister Clades
}
