% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tree.R
\name{update_tree}
\alias{update_tree}
\title{Update Phylogenetic Tree}
\usage{
update_tree(tree, species, folder)
}
\arguments{
\item{tree}{A `treedata` object representing the phylogenetic tree.}

\item{species}{A vector of species to be kept,}
}
\value{
A `treedata` object with updated metadata for sister clade groups.
}
\description{
Update Phylogenetic Tree
}
